Split PyViewerBase to h and cpp
